Create row system to group squares inside
Created a row system so that the user inputted number
creates user entered number of rows, then generates
the same number of squares within each row. Used
a nested loop to accomplish this, loop 1 creates
a row, then loop 2 creates the required number
of squares for the row. The process continues until
the number of rows and squares created equals
the user input.